requirements.txt
streamlit>=1.32.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
plotly>=5.18.0
pandas>=2.0.0
